backend/requirements.txt
fastapi==0.110.0
uvicorn==0.30.0
pymongo==4.6.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==1.10.13
